Fluentd: Attach uuid to fluentd config etc
By tying the fluent condfiguration to the release, it will be re-rendered
if the release is upgraded. This is useful in combination with [0], allowing
powerful configuration updates using helm upgrade. For example

Values:

    .Values.pod.env.fluentd.vars.OUTPUT_ENABLED: true

fluent.conf:

    ...
    {{- if .Values.pod.env.fluentd.vars.OUTPUT_ENABLED }}
    <match **>
    # Output Configuration here
    </match>
    {{- end }}

To disable this output section, issue a helm upgrade command and set the
apprpriate value to false.

    helm upgrade fluentd ./fluentd --set pod.env.fluentd.vars.OUTPUT_ENABLED=false

[0] https://review.opendev.org/#/c/726880/

Change-Id: I3dce9e5c4eaf588569e8cc3e1ea3cf3bebd0c3c5