[ceph-osd] Optimize symlink creation in udev_settle
This change optimizes the creation of symlinks in udev_settle by
only looking at the disks related to the OSD being
deployed/started and skipping the ln command for existing symlinks.
A second "udevadm settle" command is also added after the creation
of the symlinks in order to allow any logical volumes related to
new symlinks to become established.

Change-Id: I3283021fd80c8a05f0aa0c9917bb7ba0ea144303