ceph-osd: Simplify failure domain config using map
Using a subset of the characters in the hostname to determine the
failure domain is not always possible, and using overrides based on
hostnames is in some ways overkill.

This change provides a simple way to map hostnames to failure domains.
It is used only when 'failure_domain' is set other than 'host', and when
'failure_domain_by_hostname' is 'false'. Any hosts not referenced in the
map will be given the default treatment (root=default host=hostname)

Example usage:

    conf:
      storage:
        failure_domain: rack
        failure_domain_by_hostname_map:
          hostfoo: rack1
          hostbar: rack1
          hostbaz: rack2
          hostqux: rack2

Change-Id: Ia98fec8c623486f80054877e40e0753e4b939e8e