Add new python roles to playbooks
With the latest infra update, the images used no longer contain
python by default and projects are expected to use the new
ensure roles to use packages as needed.

This change adds some of the ensure roles to a few playbooks,
additional cleanup can be done using these in future changes.

Change-Id: Ie14ab297e71195d4fee070af253edf4d25ee5d27