[ceph-osd] Liveness probe success in preboot state with noup flag
OSDs fail the liveness probe if they can't make it to the 'active'
state. The noup flag keeps OSDs in the 'preboot' state, which
prevents the liveness probe from succeeding. This change adds an
additional check in the liveness probe to allow it to succeed if
the noup flag is set and OSDs are in the 'preboot' state.

Change-Id: I8df5954f7bc4ef4374e19344b6e0a9130764d60c