Remove the Elasticsearch Wait job from the chart
The elastic-cluster-wait job was meant to serve as a dependency check
for a couple of other jobs, such that when this wait job was complete
the other jobs could procede successfully. This goal can be achieved
by using our HTK init container's dependency check however.

The two jobs that waited on this wait job just need to use the
elasticsearch API, which is available once the `elasticsearch-logging`
service has endpoints.

Change-Id: I87e1c1fe3d61680a73701d48f85e5c48c11b6325