[ceph-client] Add back a new version of reweight_osds()
https://review.opendev.org/733193 removed the reweight_osds()
function from the ceph-client and weighted OSDs as they are added
in the ceph-osd chart instead. Since then some situations have
come up where OSDs were already deployed with incorrect weights
and this function is needed in order to weight them properly later
on. This new version calculates an expected weight for each OSD,
compares it to the OSD's actual weight, and makes an adjustment if
necessary.

Change-Id: I58bc16fc03b9234a08847d29aa14067bec05f1f1