Fluentd: Remove Deployment Option
This chart could deploy fluentd either as a Deployment
or a Daemonset. Both options would use the deployment-fluentd
template with various sections toggled off based on values.yaml

I'd like to know - Does anyone run this chart as a Deployment?
We can simplify the chart, and zuul gates, by changing the chart
to deploy a Daemonset specifically.

Change-Id: Ie88ceadbf5113fc60e5bb0ddef09e18fe07a192c