[ceph-osd] Allow logical disk labels to change for Ceph OSD disks
This change allows Ceph OSDs to respond to logical disk changes
and continue to function instead of failing to initialize after
such a change. For example, /dev/sdd is deployed as an OSD disk
and then subsequently becomes /dev/sde due to a hardware-related
event. This change allows the OSD to adapt and run as /dev/sde.

Change-Id: I6c22088b8d884f9dd300d026415fb126af4b41d4