[ceph-osd] Don't set CEPH_LVM_PREPARE to 0 for colocated db/wal
Due to some recent reordering of the Bluestore OSD init code, the check
for empty db and wal device strings that sets CEPH_LVM_PREPARE to 0 is
now incorrect as it checks for an existing volume group on the OSD
device to determine if an OSD previously existed on the device. That
device is now initialized prior to this check, so the check is invalid.
This change removes it.

Change-Id: I5236de171d94930e08770537663b14c2eedb0b32