lint: enable running on changed charts only
At the moment, we are using --all which means run the linting on all of
the charts.  However, the problem with using --all is that it disables
version checking which means we can't enforce version changes on Helm
charts.

This patch drops it which means the chart-testing logic will go over the
changed files and make sure that it lints those charts which have
undergone changes.

Because we use a mdoel of 1 commit per merge within Gerrit, this should
still give us the exact coverage that we need without potentially
missing any linting changes.

Change-Id: I64c7896b25c1f3daaa4f61723de8a6c722aaf3a6