Fix disappearing metacontroller CRDs on upgrade
The existing metacontroller chart conditionally only renders the CRDs if
the metacontroller does not exist.

This creates an oscillatory effect every time the chart is upgraded - if
CRDs are present, then they will be removed, and if they are absent,
they will be installed.

This change removes the metacontroller.k8s.io/v1alpha1 capabilities
check, and relies on the values.yaml option 'manifests.crds' only to
decide whether or not to render the CRDs. In an upgrade, tiller should
do the right thing based on whether the CRDs need updating.

Change-Id: I683c9e5695b7fcdddc8b6ef8622cddb96797111c