[Cinder] Updates for external ceph backend
Example override yaml file is added to indicate how to
override the manifest for configure an additional externally
managed Ceph Cinder backend.

In ceph.conf, either "mon_host" or "mon host" can be used for
the same parameter. In order not to force the user to use it one
way or the other, "mon_host" is removed from default setting.

Change-Id: I179567d77196ab2fb82d7a78e3a08efb966ed68c