Fix ks-user script case matching for domain
Some services attempt to recreate the default domain
with both the values of "default" and "Default". Since this
domain already exists when keystone is deployed, this
creates redundant API calls that only result in conflicts.

This change enables nocasematch for string checking in order
to avoid making multiple unnecessary calls to keystone.

Change-Id: I698fd420dc41eae211a511269cb021d4ab7a5bfc