[keystone] Fix fernet secret reset
This patch makes the fernet and credential secret something that gets
created only once when the deployment is first done, as when using Helm,
it's possible that it overrides it's values with an empty secret in the
runs afterwards.

By making it a hook, it will instead create it and leave an owner
reference in Helm 3 to delete it later if the release is deleted.  It
will not manage it afterwards as well.

Change-Id: I7c1c97f38877e0e54bea7fc09b37dd6f77c9dc8a