Allow openstack service list to retry in event of keystone connection issues
We've seen a few cases where the openstack service list is unable
to establish a connection with keystone thus causing the check to fail.
When this happens, an additional service is created unnecessarily.
When the addtional service is created, it tends to cause issues since
there are no endpoints asscociated with the new service.

Allow this check to retry several times.

Change-Id: I5a1985c680e90de71549177ffc3faf848a831bfa