Added detailed FiXME for ks-service script bug and code changes in [0]
On somewhat rare occasions the openstack service list call fails with
a connection aborted OSError 104 ECONNRESET. During an upgrade this failure
causes the script to think that the service it is checking for does not
exist and therefore it recreates the script. In turn this causes further
issues when other services try to use this duplicate service.
This is a temporary change in order to alliviate the issue while the root
cause is investigated.

[0] https://review.opendev.org/c/openstack/openstack-helm-infra/+/772416

Change-Id: Id0971a95eb54eca9486a9811f7ec6f603a007cbb