Add helm.sh/hook related annotation for neutron
Chart upgrading was failing due to some immutable fields in job are needed to upgrade. So, we thought using the post-install and post-upgrade
helm hook for job to force the job resource to execute after all resources are created. And as some jobs are dependent on each other i.e.
some jobs needs to run in order for helm hook to be successful. For that we used hook-weight to control resource creation order.

Change-Id: I26881324d101a986b7367d4682e9adcd07a24b13