Added post-install and post-upgrade hook for jobs.
Chart upgrading was failing due to some immutable fields are needed to upgrade before the jobs can be upgraded. For solving this issue, we
have added the helm.sh/hook annotations with post-install and post-upgrade values.
As for hook-weight annotations, we have added these to control the flow of the jobs with hook creation as the jobs are dependent. Like,
db-init jobs need to run before db-sync and so on.
helm3_hook values is added so that hook can be disabled from the values files.

Change-Id: Idfcc0479d152286ecd144502d80732094c9e43ea