Adjust Prometheus http readiness probe path from /status to /-/ready
Prometheus documentation shows that /-/ready can be used to check that
it is ready to service traffic (i.e. respond to queries) [0]. I've
witnessed cases where Prometheus's readiness probe is passing during
initial deployment using /status, which in turn triggers its helm test
to start. Said helm test then fails because /status is not a good a
reliable indicator that Prometheus is actually ready to serve traffic
and the helm test is performing actions that require it to be proprely
up and ready.

[0]: https://prometheus.io/docs/prometheus/latest/management_api/

Change-Id: Iab22d0c986d680663fbe8e84d6c0d89b03dc6428