[Update] ES helm-test script updated
This ps removes the test_api_object_creation function as the api_objects map is now more
dynamic with ability to create, delete etc.
This function throws error when it does a GET on the objects that first
needs to be created(PUT).
This function is no longer relevant with the updated create-templates
job which is more robust.

Change-Id: I9f37c86ae9ca4bf32c417880926b6a3c3e78cb8a