Make Ceph pool init job consistent with helm test
The current pool init job only allows the finding of PGs in the
"peering" or "activating" (or active) states, but it should also
allow the other possible states that can occur while the PG
autoscaler is running ("unknown" and "creating" and "recover").
The helm test is already allowing these states, so the pool init
job is being changed to also allow them to be consistent.

Change-Id: Ib2c19a459c6a30988e3348f8d073413ed687f98b