[ceph-client] Add pool rename support for Ceph pools
A new value "rename" has been added to the Ceph pool spec to allow
pools to be renamed in a brownfield deployment. For greenfield the
pool will be created and renamed in a single deployment step, and
for a brownfield deployment in which the pool has already been
renamed previously no changes will be made to pool names.

Change-Id: I3fba88d2f94e1c7102af91f18343346a72872fde