Remove misleading section from upgrade docs
The openstack-helm docs currently state that the openstack-helm
dev team will work with Helm with regards to facilitating job upgrades.

This is misleading in that we do not directly contribute to Helm and
currently provide methods for charts to run jobs for an upgrade
for instances where images are updated, the job is deleted and
re-ran.

Change-Id: If04367b6563ed36c3b3cde7a9cd4425b6795505f