[ceph-client] Add pool delete support for Ceph pools
Two new values, "delete" and "delete_all_pool_data," have been
added to the Ceph pool spec to allow existing pools to be deleted
in a brownfield deployment. For deployments where a pool does not
exist, either for greenfield or because it has been deleted
previously, the pool will be created and then deleted in a single
step.

Change-Id: Ic22acf02ae2e02e03b834e187d8a6a1fa58249e7