[ceph-rgw] Fix a bug in placement target deletion for new targets
A deployment that specifies a placement target with "delete: true"
should delete that placement target if it exists. For a clean
deployment the expectation is that the placement target should be
created and immediately deleted; however, the check for existence
happens before its creation and the delete doesn't execute as a
result. This change adds a recheck for existence immediately after
creation to remedy that.

Change-Id: I26f7fa79c5c851070e94af758d0a0438aa7efa52